requirements.txt
aioboto3==10.0.0
aiobotocore==2.3.4
aiofiles==0.8.0
aiogram==2.22.1
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
aiosqlite==0.17.0
appdirs==1.4.4
asqlite3==0.0.1
async-timeout==4.0.2
attrs==22.1.0
audio-to-midi==2020.7
autocorrect==2.6.1
Babel==2.9.1
beautifulsoup4==4.11.1
boto3==1.21.21
botocore==1.24.21
bottle==0.12.23
bs4==0.0.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
colorama==0.4.5
conan==1.51.2
dill==0.3.5.1
et-xmlfile==1.1.0
fasteners==0.17.3
frozenlist==1.3.1
funcy==1.17
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.12.0
jamspell==0.0.12
Jinja2==3.1.2
jmespath==1.0.1
loguru==0.6.0
lxml==4.9.1
MarkupSafe==2.1.1
mtranslate==1.8
multidict==6.0.2
aiofiles==0.8.0
aiohttp==3.8.1
funcy==1.17
Pillow==9.2.0
Pyrogram==2.0.51
TgCrypto==1.2.3
typing_extensions==4.3.0
